[ESP8266] Fixed compiler warnings

+++ src/modules/ESP8266.cpp
@@ -189,7 +189,7 @@ size_t ESP8266::receive(uint8_t* data, size_t len, uint32_t timeout) {
 uint16_t ESP8266::getNumBytes(uint32_t timeout, size_t minBytes) {
   // wait for available data
   uint32_t start = millis();
-  while(_mod->ModuleSerial->available() < minBytes) {
+  while(_mod->ModuleSerial->available() < (int16_t)minBytes) {
     if(millis() - start >= timeout) {
       return(0);
     }
